move cloud library into pkg
