+++ pkg/cloud/aws/client/mock/client_generated.go
