+++ pkg/cloud/aws/logging/logging.go
