+++ pkg/cloud/aws/providerconfig/doc.go
@@ -12,8 +12,8 @@
 // limitations under the License.
 
 // +k8s:deepcopy-gen=package,register
-// +k8s:conversion-gen=sigs.k8s.io/cluster-api-provider-aws/cloud/aws/providerconfig
+// +k8s:conversion-gen=sigs.k8s.io/cluster-api-provider-aws/pkg/cloud/aws/providerconfig
 // +k8s:openapi-gen=true
 // +k8s:defaulter-gen=TypeMeta
 
-package v1alpha1
+package providerconfig
